Fix types for opening times
