+++ common/views/components/PageLayout/PageLayout.tsx
@@ -11,7 +11,6 @@ import NewsletterPromo from '../NewsletterPromo/NewsletterPromo';
 import Footer from '../Footer/Footer';
 import PopupDialogContext from '../PopupDialogContext/PopupDialogContext';
 import PopupDialog from '../PopupDialog/PopupDialog';
-import OpeningTimesContext from '../OpeningTimesContext/OpeningTimesContext';
 import Space from '../styled/Space';
 import GlobalAlertContext from '../GlobalAlertContext/GlobalAlertContext';
 import { museumLd, libraryLd } from '../../../utils/json-ld';
@@ -19,6 +18,7 @@ import { collectionVenueId } from '../../../services/prismic/hardcoded-id';
 import {
   getParseCollectionVenueById,
   openingHoursToOpeningHoursSpecification,
+  parseCollectionVenues,
 } from '../../../services/prismic/opening-times';
 import { wellcomeCollectionGallery } from '../../../model/organization';
 import GlobalContextProvider, {
@@ -28,7 +28,7 @@ import GlobalInfoBarContext, {
   GlobalInfoBarContextProvider,
 } from '../GlobalInfoBarContext/GlobalInfoBarContext';
 import ApiToolbar from '../ApiToolbar/ApiToolbar';
-import { useToggles } from '../../../server-data/Context';
+import { usePrismicData, useToggles } from '../../../server-data/Context';
 import useHotjar from '../../../hooks/useHotjar';
 
 export type SiteSection =
@@ -111,7 +111,8 @@ const PageLayoutComponent: FunctionComponent<ComponentProps> = ({
   const absoluteUrl = `https://wellcomecollection.org${urlString}`;
   const globalAlert = useContext(GlobalAlertContext);
   const popupDialog = useContext(PopupDialogContext);
-  const openingTimes = useContext(OpeningTimesContext);
+  const prismicData = usePrismicData();
+  const openingTimes = parseCollectionVenues(prismicData.openingTimes);
   const galleries =
     openingTimes &&
     getParseCollectionVenueById(openingTimes, collectionVenueId.galleries.id);
