+++ common/views/components/VenueClosedPeriods/VenueClosedPeriods.tsx
@@ -1,24 +1,25 @@
 import { Venue } from '@weco/common/model/opening-hours';
-import { useContext } from 'react';
 import {
   backfillExceptionalVenueDays,
   getExceptionalOpeningPeriods,
   getExceptionalClosedDays,
   groupConsecutiveDays,
   convertJsonDateStringsToMoment,
+  parseCollectionVenues,
 } from '../../../services/prismic/opening-times';
 import { formatDayDate } from '@weco/common/utils/format-date';
-import OpeningTimesContext from '@weco/common/views/components/OpeningTimesContext/OpeningTimesContext';
 import {
   collectionVenueId,
   getNameFromCollectionVenue,
 } from '@weco/common/services/prismic/hardcoded-id';
+import { usePrismicData } from '../../../server-data/Context';
 type Props = {
   venue: Venue;
 };
 
 const VenueClosedPeriods = ({ venue }: Props) => {
-  const openingTimes = useContext(OpeningTimesContext);
+  const prismicData = usePrismicData();
+  const openingTimes = parseCollectionVenues(prismicData.openingTimes);
   const exceptionalPeriods = getExceptionalOpeningPeriods(openingTimes);
   const backfilledExceptionalPeriods = backfillExceptionalVenueDays(
     convertJsonDateStringsToMoment(venue),
